requirements.txt
django
office_word_count
python-docx
requests
textract
